feat: insight-ui
