--- src/_packages/@lunit/insight-ui/colors/emphasize.ts
@@ -1,6 +1,0 @@
-import { PaletteColor } from '@material-ui/core/styles/createPalette';
-
-export function emphasize(color: PaletteColor, elevation: number): PaletteColor {
-  // TODO
-  return color;
-}
