+++ src/_packages/@lunit/insight-ui/components/Dialog.tsx
@@ -1,14 +1,13 @@
 import { Dialog as MuiDialog, DialogProps } from '@material-ui/core';
 import { ComponentType } from 'react';
 import styled from 'styled-components';
-import { lighten } from '../theme/lighten';
 
 export { DialogActions, DialogContent, DialogContentText, DialogTitle } from '@material-ui/core';
 
 export const Dialog: ComponentType<DialogProps> = styled(MuiDialog)`
   && {
     .MuiDialog-paper {
-      ${lighten};
+      // TODO {lighten};
 
       background-color: #ffffff;
