+++ src/_packages/@lunit/insight-ui/components/buttons/Button.tsx
@@ -1,13 +1,28 @@
-import { ButtonBase, ButtonBaseProps } from '@material-ui/core';
+import { ButtonBase, ButtonBaseClassKey, ButtonBaseProps } from '@material-ui/core';
 import { PaletteColor } from '@material-ui/core/styles/createPalette';
-import React from 'react';
-import { useButtonStyle } from './useButtonStyle';
+import { DefaultTheme, Styles, withStyles } from '@material-ui/styles';
+import React, { ComponentType, CSSProperties } from 'react';
+import { buttonStyle } from './buttonStyle';
 
-export interface ButtonProps extends Omit<ButtonBaseProps, 'color'> {
-  color?: PaletteColor;
+export interface LunitControl {
+  color?: { root: PaletteColor };
+  fit?: 'form' | 'sidebar';
 }
 
-export function Button({ color, ...buttonProps }: ButtonProps) {
-  const classes = useButtonStyle({ color });
-  return <ButtonBase {...buttonProps} classes={classes} />;
+export interface LunitButtonProps extends Omit<ButtonBaseProps, 'color'>, LunitControl {}
+
+export function Button2({ color, fit, style, classes, ...buttonProps }: LunitButtonProps) {
+  const buttonStyle: CSSProperties | undefined = color?.root
+    ? {
+        ...style,
+        '--button-background-color': color.root.main,
+        '--button-color': color.root.contrastText,
+      }
+    : style;
+
+  return <ButtonBase {...buttonProps} style={buttonStyle} />;
 }
+
+export const LunitButton: ComponentType<LunitButtonProps> = withStyles<
+  Styles<DefaultTheme, LunitButtonProps, ButtonBaseClassKey>
+>(buttonStyle, { name: 'LunitButton' })(Button2);
