+++ src/_packages/@lunit/insight-ui/components/buttons/LinkButton.tsx
@@ -1,13 +1,13 @@
 import * as H from 'history';
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { Button, ButtonProps } from './Button';
+import { LunitButton, LunitButtonProps } from './Button';
 
-export interface LinkButtonProps extends ButtonProps {
+export interface LinkButtonProps extends LunitButtonProps {
   to: H.LocationDescriptor | ((location: H.Location) => H.LocationDescriptor);
 }
 
 export const LinkButton = ({ to, ...buttonProps }: LinkButtonProps) => {
   // @ts-ignore: see https://github.com/mui-org/material-ui/issues/7877
-  return <Button {...buttonProps} component={Link} to={to} />;
+  return <LunitButton {...buttonProps} component={Link} to={to} />;
 };
