--- src/_packages/@lunit/insight-ui/components/buttons/useButtonStyle.ts
@@ -1,11 +1,0 @@
-import { ButtonBaseClassKey } from '@material-ui/core';
-import { PaletteColor } from '@material-ui/core/styles/createPalette';
-import { DefaultTheme, makeStyles } from '@material-ui/styles';
-
-export const useButtonStyle = makeStyles<DefaultTheme, { color?: PaletteColor }, ButtonBaseClassKey>((theme) => {
-  return {
-    root: {},
-    disabled: {},
-    focusVisible: {},
-  };
-});
