+++ src/_packages/@lunit/insight-ui/index.ts
@@ -1,31 +1,27 @@
+export * from './components/buttons/Button';
+export * from './components/buttons/LinkButton';
+export * from './components/buttons/ToggleButton';
+
 export * from './components/containers/ScreenFitContainer';
 export * from './components/containers/ScrollContainer';
 
 export * from './components/layout/useWindowSize';
 export * from './components/layout/useTouchDeviceLayoutFixer';
 
-export * from './components/buttons/LinkButton';
-//export * from './components/buttons/Button'
-export * from './components/buttons/ToggleButton';
+export * from './theme/createInsightTheme';
+export * from './theme/globalStyle';
+export * from './theme/internal';
+export * from './theme/lunitDark';
+export * from './theme/ThemeProvider';
+export * from './theme/types';
 
-export * from './colors/internal';
-export * from './colors/lunitDark';
-//export * from './colors/lunitLight';
-
-export * from './theme/lighten';
-export * from './theme/style';
-export * from './theme/theme';
-
-export * from './theme/lunitDarkTheme';
-
-export * from './components/theme/ThemeProvider';
-
+// TODO Remove Legacy
 export * from './components/Button';
+export * from './components/ButtonLayout';
 export * from './components/Panel';
 export * from './components/PanelDescription';
 export * from './components/PanelToolbar';
 export * from './components/PanelToolbarBadge';
-export * from './components/ButtonLayout';
 export * from './components/Slider';
 export * from './components/Tooltip';
 export * from './components/Switch';
