+++ src/_packages/@lunit/insight-ui/theme/ThemeProvider.tsx
@@ -1,11 +1,11 @@
-import { Theme } from '@lunit/insight-ui';
+import { InsightTheme } from '@lunit/insight-ui/theme/types';
 import { ThemeProvider as MuiThemeProvider, useTheme } from '@material-ui/styles';
 import React, { ReactNode, useMemo } from 'react';
 import { ThemeProvider as StyledComponentsThemeProvider } from 'styled-components';
 
 export interface ThemeProviderProps {
   children: ReactNode;
-  theme: Theme;
+  theme: InsightTheme;
 }
 
 export function ThemeProvider({ children, theme }: ThemeProviderProps) {
