+++ src/_packages/@lunit/insight-ui/theme/createInsightTheme.ts
@@ -1,0 +1,5 @@
+import { InsightTheme, InsightThemeOptions } from './types';
+
+export function createInsightTheme(option: InsightThemeOptions): InsightTheme {
+  throw new Error();
+}
