--- src/_packages/@lunit/insight-ui/theme/lighten.ts
@@ -1,41 +1,0 @@
-import { createMuiTheme, Theme as MuiTheme } from '@material-ui/core/styles';
-import { css } from 'styled-components';
-
-export const lighten = css`
-  color: #242e3e;
-
-  --button-background-color: #00a4c8;
-  --button-label-color: rgba(255, 255, 255, 0.8);
-  --button-background-color-hover: #008cb0;
-  --button-label-color-hover: rgba(255, 255, 255, 1);
-  --button-background-color-selected: #007294;
-  --button-label-color-selected: rgba(255, 255, 255, 1);
-  --button-background-color-disabled: rgba(0, 164, 200, 0.65);
-  --button-label-color-disabled: rgba(255, 255, 255, 0.2);
-
-  --slider-rail-color: rgba(86, 81, 136, 0.4);
-  --slider-thumb-color: #6b6b9b;
-  --slider-track-color: rgba(86, 81, 136, 0.6);
-  --slider-value-label-color: #ffffff;
-`;
-
-export const lightenTheme: MuiTheme = createMuiTheme({
-  palette: {
-    type: 'light',
-  },
-  typography: {
-    fontFamily: 'proximanova,noto_sanslight,sans-serif',
-    fontWeightLight: 200,
-    fontWeightRegular: 400,
-    fontWeightMedium: 600,
-    fontWeightBold: 700,
-  },
-  overrides: {
-    MuiTooltip: {
-      tooltip: {
-        fontSize: 13,
-        backgroundColor: 'rgba(0, 0, 0, 0.5)',
-      },
-    },
-  },
-});
