--- src/_packages/@lunit/insight-ui/theme/theme.ts
@@ -1,9 +1,0 @@
-import { Theme as MuiTheme, ThemeOptions as MuiThemeOptions } from '@material-ui/core/styles';
-
-export interface UserTheme {}
-
-export interface UserThemeOptions {}
-
-export interface Theme extends MuiTheme, UserTheme {}
-
-export interface ThemeOptions extends MuiThemeOptions, UserThemeOptions {}
